rename: article auth api 파일 이름 수정
