+++ pages/boards/index.tsx
@@ -1,7 +1,7 @@
 import AllArticlesBoard from '@components/board/AllArticlesBoard/AllArticlesBoard';
 import BestArticlesSection from '@components/board/BestArticlesSection/BestArticlesSection';
 import BasicLayout from '@components/layout/BasicLayout/BasicLayout';
-import { getArticles } from '@lib/api/articleApi';
+import { getArticles } from '@lib/api/article.api';
 import { Article } from '@type/ArticleTypes';
 import { error } from 'console';
 import { GetServerSideProps } from 'next';
